chore: move slug api to app router
